Move enabling of service and configuring of Nessus agent
Move these two tasks inside the when not nessus_directory.stat.exists
block so that there is a check in place to ensure the configuring of
the Nessus agent is idempotent.  The enabling of the Nessus agent
service must move too because it is required by the nessus agent
configuration task.