Install cisagov/ansible-role-cdm-certificates in the role tasks
This is done instead of relying on importing via the metadata
dependencies.  This allows us to set a role variable for the CDM
certificates role using the role variables for this role.

Note that roles listed in meta/main.yml are installed before the
current role, whereas meta/requirements.yml are merely made available
to the current role.  This is a key difference that we exploit.